add captcha and timer

+++ src/services/socket.js
@@ -2,11 +2,13 @@ const io = require('socket.io')
 const debug = require('debug')('prisoners:socket')
 
 const dilemmaService = require('./dilemma')
+const captchaService = require('./captcha')
 
 let _clientId = 0
 
 const socketService = {
   _clients: {},
+  _connections: {},
 
   init: (server) => {
     const socket = io(server)
@@ -18,12 +20,25 @@ const socketService = {
 
   addClient: (id, client) => {
     debug('add client', id)
+
+    const address = client.handshake.address
+    socketService._connections[address] = socketService._connections[address] || 0
+    if (socketService._connections[address] >= 10) {
+      debug('limiting', address)
+      client.disconnect()
+      return
+    }
+
+    socketService._connections[address]++
     socketService._clients[id] = client
 
-    client.on('reset', () => {
+    client.on('reset', async (token) => {
       dilemmaService.removePlayer(id)
-      const dilemma = dilemmaService.newPlayer(id)
-      socketService.emitDilemma(dilemma)
+      const verified = await captchaService.verify(client.handshake.address, token)
+      if (verified) {
+        const dilemma = dilemmaService.newPlayer(id)
+        socketService.emitDilemma(dilemma)
+      }
     })
 
     client.on('message', message => {
@@ -39,6 +54,7 @@ const socketService = {
       debug('disconnect', id)
       const dilemma = dilemmaService.removePlayer(id)
       delete socketService._clients[id]
+      socketService._connections[address]--
       socketService.emitDilemma(dilemma)
     })
   },
